requirements.txt
# Core API Framework
fastapi==0.110.0
uvicorn==0.29.0
python-dotenv==1.0.1
pydantic==2.6.1
python-multipart==0.0.9

# AI and NLP Libraries
groq==0.4.1
openai==1.12.0
requests==2.31.0

# Financial Data APIs
yfinance==0.2.37
pandas==2.2.1
numpy==1.26.4

# Real-time Stock Data APIs
aiohttp==3.9.3
upstox-python-sdk==1.2.0

# News and Market Data
newsapi-python==0.2.7

# Google Cloud APIs (Python 3.12 compatible versions)
google-cloud-aiplatform>=1.38.0
google-cloud-vision>=3.4.0
google-cloud-bigquery>=3.11.0
google-cloud-dialogflow>=2.20.0
google-cloud-storage>=2.10.0
google-auth>=2.20.0
google-auth-oauthlib>=1.0.0

# Azure APIs (Python 3.12 compatible versions)
azure-cognitiveservices-language-textanalytics>=0.2.2
azure-identity>=1.15.0

# Database and Storage
redis==5.0.2

# API Clients
httpx==0.27.0
aiohttp==3.9.3

# Utilities
typing-extensions==4.9.0

# Web scraping libraries
beautifulsoup4==4.12.2
lxml==5.1.0
feedparser==6.0.11
html5lib==1.1
python-dateutil==2.9.0

# Data refresh scheduler dependencies
apscheduler==3.11.0
